update to categorization cmd
